+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,7 @@ Release History
 Upcoming
 ++++++
 * 'az containerapp job execution show/list': improve table output format
+* 'az containerapp create/update': --yaml support properties for api-version 2023-04-01-preview (e.g. subPath, mountOptions)
 
 0.3.33
 ++++++